requirements.txt
Django~=4.0.6
requests~=2.28.1
djangorestframework~=3.13.1